# nutriguard-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NutriGuard Backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/scan-history": {
      "get": {
        "tags": [
          "scan-history"
        ],
        "summary": "List Scan History",
        "description": "API Contract 6.10",
        "operationId": "list_scan_history_api_v1_scan_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ScanHistoryOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scan-history"
        ],
        "summary": "Clear Scan History",
        "description": "API Contract 6.11",
        "operationId": "clear_scan_history_api_v1_scan_history_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_ScanHistoryOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScanHistoryOut"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          },
          "totalItems": {
            "type": "integer",
            "title": "Totalitems"
          },
          "totalPages": {
            "type": "integer",
            "title": "Totalpages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "totalItems",
          "totalPages"
        ],
        "title": "Page[ScanHistoryOut]"
      },
...
      "ScanHistoryOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "barcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          },
          "productName": {
            "type": "string",
            "title": "Productname"
          },
          "brand": {
            "type": "string",
            "title": "Brand"
          },
          "healthScore": {
            "type": "integer",
            "title": "Healthscore"
          },
          "scannedAt": {
            "type": "integer",
            "title": "Scannedat"
          },
          "scanType": {
            "$ref": "#/components/schemas/ScanType"
          }
        },
        "type": "object",
        "required": [
          "id",
          "productName",
          "brand",
          "healthScore",
          "scannedAt",
          "scanType"
        ],
        "title": "ScanHistoryOut",
        "description": "Mirrors com.example.data.model.ScanHistoryEntity (API Contract 5.6)."
      },
      "ScanType": {
        "type": "string",
        "enum": [
          "BARCODE",
          "OCR_LABEL",
          "MANUAL_INPUT"
        ],
        "title": "ScanType",
        "description": "Mirrors the scanType string values used in ScanHistoryEntity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}